+++ .gitignore
@@ -1,4 +1,4 @@
-[Cc]ache/*
+*/[Cc]ache/*
 *.m3u
 
 ## Ignore Visual Studio temporary files, build results, and
