+++ IptvPlaylistAggregator/Communication/IFileDownloader.cs
