+++ IptvPlaylistAggregator/Configuration/ApplicationSettings.cs
