+++ IptvPlaylistAggregator/Configuration/CacheSettings.cs
