+++ IptvPlaylistAggregator/Configuration/DataStoreSettings.cs
