+++ IptvPlaylistAggregator/DataAccess/DataObjects/PlaylistProviderEntity.cs
@@ -1,0 +1,22 @@
+using NuciDAL.DataObjects;
+
+namespace IptvPlaylistAggregator.DataAccess.DataObjects
+{
+    public class PlaylistProviderEntity : EntityBase
+    {
+        public bool IsEnabled { get; set; }
+
+        public int Priority { get; set; }
+
+        public string Name { get; set; }
+
+        public string UrlFormat { get; set; }
+
+        public string ChannelNameOverride { get; set; }
+
+        public PlaylistProviderEntity()
+        {
+            Priority = int.MaxValue;
+        }
+    }
+}
