+++ IptvPlaylistAggregator/Logging/MyLogInfoKey.cs
