+++ IptvPlaylistAggregator/Logging/MyOperation.cs
