+++ IptvPlaylistAggregator/Service/IMediaSourceChecker.cs
