+++ IptvPlaylistAggregator/Service/IPlaylistAggregator.cs
