+++ IptvPlaylistAggregator/Service/IPlaylistFetcher.cs
