+++ IptvPlaylistAggregator/Service/IPlaylistFileBuilder.cs
