+++ IptvPlaylistAggregator/Service/Mapping/ChannelDefinitionMapping.cs
