+++ IptvPlaylistAggregator/Service/Mapping/GroupMapping.cs
