+++ IptvPlaylistAggregator/Service/Mapping/PlaylistProviderMapping.cs
