+++ IptvPlaylistAggregator/Service/Models/Channel.cs
