+++ IptvPlaylistAggregator/Service/Models/ChannelDefinition.cs
