+++ IptvPlaylistAggregator/Service/Models/Group.cs
