+++ IptvPlaylistAggregator/Service/Models/MediaStreamStatus.cs
