+++ IptvPlaylistAggregator/Service/Models/Playlist.cs
