+++ IptvPlaylistAggregator/Service/Models/PlaylistProvider.cs
