+++ IptvPlaylistAggregator/Service/PlaylistFileBuilder.cs
