+++ IptvPlaylistAggregator/appsettings.json
